requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
python-dotenv==1.0.1
yookassa==2.3.5
qrcode==7.4.2
colorama==0.4.6
flask==3.0.2
Pillow==10.2.0
